vagrant: updated fedora box to version 30
An update to Fedora 31 is possible but right now "too much" to add in the bit
of time I have to update the project right now.

In Fedora 31 Docker doesn't work out of the box right now due to cgroups
v2 being set to unified, see
https://bugzilla.redhat.com/show_bug.cgi?id=1746355 for more
information.

Add the update-alternatives for iptables which is needed as described
here: https://kubernetes.io/docs/setup/production-environment/tools/kubeadm/#ensure-iptables-tooling-does-not-use-the-nftables-backend

Signed-off-by: Alexander Trost <galexrt@googlemail.com>