make: improve make status and KUBERNETES_VERSION
This makes the `make status` remove any empty lines from the vagrant
status call happening.
The KUBERNETES_VERSION env var now allows for a preceding `v` to be
given. It will be removed in the underlying "code". This makes it easier
to just curl out to kubernetes `stable.txt` file for the latest K8S
version.

Signed-off-by: Alexander Trost <galexrt@googlemail.com>