Allow creating single-node master-only clusters
Allow creating single-node master-only clusters.
`seq 1 $(NODE_COUNT)` will naturally count backwards if
NODE_COUNT==0, and it will return an error if the form
`seq <start> <end> <increment>` is given when the
increment==1 and NODE_COUNT==0. Change for loops using
`seq` to instead use the C-like for loop form which
behaves as expected and does not cause errors if
NODE_COUNT is less than 1.

Signed-off-by: Blaine Gardner <b.blaine.gardner@gmail.com>